requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
google-cloud-aiplatform
python-dotenv
pydantic-settings
requests
jinja2
